새로고침 시 페이지 유지, url 파라미터 명시

+++ src/main/java/com/adam9e96/WordLOL/mapper/WordMapper.java
@@ -19,5 +19,5 @@ public interface WordMapper {
 
     Long countAllWords();
 
-    List<EnglishWord> findByWordBookId(int wordBookId);
+    List<EnglishWord> findByWordBookId(Long wordBookId);
 }
